added (int)

+++ src/Imper86/Core/DateTime.php
@@ -22,7 +22,7 @@ class DateTime extends \DateTime
     public function __construct($time = 'now', DateTimeZone $timezone = null)
     {
         if (substr($time, 0, 1) == '@') {
-            $time = date('Y-m-d H:i:s', substr($time, 1));
+            $time = date('Y-m-d H:i:s', (int)substr($time, 1));
         }
 
         parent::__construct($time, $timezone);
